Fix batch quality: dequant Q4L to fp32 (not fp16)
Root cause: dequanting to fp16 truncates the per-block float32 scale
factor, causing 1-3% error per layer that compounds across 28 layers
into gibberish output.

The dp4a single-sequence path doesn't have this issue because it
applies the float32 scale ONCE per block at the end (not per element).

Fix: dequant to fp32, use cuBLAS with CUDA_R_32F weight input and
CUBLAS_COMPUTE_32F accumulation. Memory: 12MB scratch (was 6MB).

+++ engine/model.h
@@ -193,8 +193,8 @@ struct BatchState {
     half* kv_keys[qwen3::NUM_LAYERS];
     half* kv_values[qwen3::NUM_LAYERS];
 
-    // Q4L dequant scratch (largest projection = INTERMEDIATE_SIZE * HIDDEN_SIZE)
-    half* dequant_scratch;
+    // Q4L dequant scratch in fp32 (largest projection = INTERMEDIATE_SIZE * HIDDEN_SIZE)
+    float* dequant_scratch;
 
     // Per-sequence state
     int* h_positions;   // host (G,)
